+++ src/main/java/com/elmakers/mine/bukkit/protection/TownyAPI.java
@@ -2,9 +2,11 @@ package com.elmakers.mine.bukkit.protection;
 
 import com.palmergames.bukkit.towny.Towny;
 import com.palmergames.bukkit.towny.TownySettings;
+import com.palmergames.bukkit.towny.db.TownyDataSource;
 import com.palmergames.bukkit.towny.exceptions.NotRegisteredException;
 import com.palmergames.bukkit.towny.exceptions.TownyException;
 import com.palmergames.bukkit.towny.object.Coord;
+import com.palmergames.bukkit.towny.object.Resident;
 import com.palmergames.bukkit.towny.object.Town;
 import com.palmergames.bukkit.towny.object.TownBlock;
 import com.palmergames.bukkit.towny.object.TownBlockType;
@@ -14,23 +16,35 @@ import com.palmergames.bukkit.towny.object.TownyWorld;
 import com.palmergames.bukkit.towny.object.WorldCoord;
 import com.palmergames.bukkit.towny.utils.CombatUtil;
 import com.palmergames.bukkit.towny.utils.PlayerCacheUtil;
+import org.bukkit.Bukkit;
 import org.bukkit.Location;
 import org.bukkit.block.Block;
 import org.bukkit.entity.Entity;
 import org.bukkit.entity.Player;
 import org.bukkit.plugin.Plugin;
 
+import java.lang.reflect.Method;
+import java.util.UUID;
+import java.util.logging.Level;
+
 public class TownyAPI
 {
     private final Towny towny;
     private final TownyManager controller;
+    private Method getPlayerByNameMethod;
+    private Method getPlayerByUUIDMethod;
 
-    public TownyAPI(TownyManager manager, Plugin plugin) throws IllegalArgumentException {
+    public TownyAPI(TownyManager manager, Plugin plugin) throws IllegalArgumentException, NoSuchMethodException {
         this.controller = manager;
         if (!(plugin instanceof Towny)) {
             throw new IllegalArgumentException("Towny plugin not an instance of Towny class");
         }
         towny = (Towny) plugin;
+        try {
+            getPlayerByNameMethod = TownyDataSource.class.getMethod("getResident", String.class);
+        } catch (Exception ex) {
+            getPlayerByUUIDMethod = TownyDataSource.class.getMethod("getResident", UUID.class);
+        }
     }
 
     public boolean isPVPAllowed(Location location) {
@@ -99,5 +113,50 @@ public class TownyAPI
         }
 
         return true;
+    }
+
+    protected Resident getResident(Player player) {
+        try {
+            if (getPlayerByNameMethod != null) {
+                return (Resident) getPlayerByNameMethod.invoke(TownyUniverse.getDataSource(), player.getName());
+            }
+            if (getPlayerByUUIDMethod != null) {
+                return (Resident) getPlayerByUUIDMethod.invoke(TownyUniverse.getDataSource(), player.getUniqueId());
+            }
+        } catch (Exception ex) {
+            if (!(ex instanceof TownyException)) {
+                Bukkit.getLogger().log(Level.WARNING, "Error getting Towny Resident", ex);
+            }
+        }
+        return null;
+    }
+
+    public Location getTownLocation(Player player) {
+        if (towny == null || player == null) {
+            return null;
+        }
+
+        Resident resident = getResident(player);
+        if (resident == null) {
+            return null;
+        }
+
+        Town town = null;
+        try {
+            town = resident.getTown();
+        } catch (Exception ex) {
+        }
+
+        if (town == null) {
+            return null;
+        }
+
+        Location location = null;
+        try {
+            location = town.getSpawn();
+        } catch (Exception ex) {
+
+        }
+        return location;
     }
 }